event handeling - send event from event processor to webapp handler

+++ ThermAlarmSln/ThermAlarm.Common/SensorsProcessing.cs
@@ -4,9 +4,11 @@ using System.Text;
 
 namespace ThermAlarm.Common
 {
+    
     public static class SensorsProcessing
     {
-        private static bool thermWarmObj(int[] thermValues)
+        
+        private static bool thermWarmObj(float[] thermValues)
         {
             //TODO - make logic better.. this is really naive approch - third of pixels above 30
             int countWarm = 0;
@@ -18,7 +20,7 @@ namespace ThermAlarm.Common
             return (countWarm > 20);
         }
 
-        public static bool shouldAlarm(int pirValue, int[] thermValues)
+        public static bool shouldAlarm(int pirValue, float[] thermValues)
         {
             if (pirValue > 0)
             {
